fix docker env

+++ docker-compose-dev.sh
@@ -5,15 +5,30 @@ java --version
 echo '### Gradle version ###'
 gradle --version
 
-docker-compose down
-docker stop $(docker ps -a -q)
-docker rm $(docker ps -a -q)
-docker rmi -f $(docker images | grep 'niffler')
+front=""
+front_image=""
+if [[ "$1" = "gql" ]]; then
+  front="./niffler-frontend-gql/";
+  front_image="dtuchs/niffler-frontend-gql:latest";
+else
+  front="./niffler-frontend/";
+  front_image="dtuchs/niffler-frontend:latest";
+fi
 
-if [[ $1 = "gql" ]]; then a="$c"; else a="$d"; fi
+FRONT_IMAGE="$front_image" docker-compose down
 
-var front
-if [[ "$1" = "gql" ]]; then front="./niffler-frontend-gql/"; else front="./niffler-frontend/"; fi
+docker_containers="$(docker ps -a -q)"
+docker_images="$(docker images --format '{{.Repository}}:{{.Tag}}' | grep 'niffler')"
+
+if [ ! -z "$docker_containers" ]; then
+  echo "### Stop containers: $docker_containers ###"
+  docker stop $(docker ps -a -q)
+  docker rm $(docker ps -a -q)
+fi
+if [ ! -z "$docker_images" ]; then
+  echo "### Remove images: $docker_images ###"
+  docker rmi $(docker images --format '{{.Repository}}:{{.Tag}}' | grep 'niffler')
+fi
 
 if [ "$1" = "push" ] || [ "$2" = "push" ]; then
   echo "### Build & push images (front: $front) ###"
@@ -29,5 +44,5 @@ fi
 
 cd ../
 docker images
-docker-compose up -d
+FRONT_IMAGE="$front_image" docker-compose up -d
 docker ps -a
