+++ docker-compose-e2e.sh
@@ -7,31 +7,48 @@ java --version
 echo '### Gradle version ###'
 gradle --version
 
-docker-compose -f docker-compose.test.yml down
-docker stop $(docker ps -a -q)
-docker rm $(docker ps -a -q)
-docker rmi -f $(docker images | grep 'niffler')
+front=""
+front_image=""
+docker_arch=""
+if [[ "$1" = "gql" ]]; then
+  front="./niffler-frontend-gql/";
+  front_image="dtuchs/niffler-frontend-gql-test:latest";
+else
+  front="./niffler-frontend/";
+  front_image="dtuchs/niffler-frontend-test:latest";
+fi
+
+ARCH="$docker_arch" FRONT_IMAGE="$front_image" docker-compose -f docker-compose.test.yml down
+
+docker_containers="$(docker ps -a -q)"
+docker_images="$(docker images --format '{{.Repository}}:{{.Tag}}' | grep 'niffler')"
+
+if [ ! -z "$docker_containers" ]; then
+  echo "### Stop containers: $docker_containers ###"
+  docker stop $(docker ps -a -q)
+  docker rm $(docker ps -a -q)
+fi
+if [ ! -z "$docker_images" ]; then
+  echo "### Remove images: $docker_images ###"
+  docker rmi $(docker images --format '{{.Repository}}:{{.Tag}}' | grep 'niffler')
+fi
 
 ARCH=$(uname -m)
 
 bash ./gradlew clean build dockerTagLatest -x :niffler-e-2-e-tests:test
 
-var DOCKER_ARCH
 if [ "$ARCH" = "arm64" ] || [ "$ARCH" = "aarch64" ]; then
-  DOCKER_ARCH="linux/arm64/v8"
+  docker_arch="linux/arm64/v8"
   docker build --build-arg DOCKER=arm64v8/eclipse-temurin:19-jdk -t "${IMAGE_NAME}":"${VERSION}" -t "${IMAGE_NAME}":latest -f ./niffler-e-2-e-tests/Dockerfile .
 else
-  DOCKER_ARCH="linux/amd64"
+  docker_arch="linux/amd64"
   docker build --build-arg DOCKER=eclipse-temurin:19-jdk -t "${IMAGE_NAME}":"${VERSION}" -t "${IMAGE_NAME}":latest -f ./niffler-e-2-e-tests/Dockerfile .
 fi
-
-var front
-if [[ "$1" = "gql" ]]; then front="./niffler-frontend-gql/"; else front="./niffler-frontend/"; fi
 
 cd "$front" || exit
 bash ./docker-build.sh test
 cd ../ || exit
 docker pull selenoid/vnc_chrome:110.0
 docker images
-ARCH="$DOCKER_ARCH" docker-compose -f docker-compose.test.yml up -d
+ARCH="$docker_arch" FRONT_IMAGE="$front_image" docker-compose -f docker-compose.test.yml up -d
 docker ps -a
